feat: created ListOrderModal component
In the ListOrderModal component you can choose between 'Alphabetical', 'Date', 'Custom' order, and if you click on the currently active order, it will sort the lists from ascending to descending order, and vice versa.

Note: yet the selected orders are still not stored! Maybe it's a good idea to store it in the data/settings.json.

Minor changes in the BottomSheet component. I dropped the snap indexes so now all input fields are displayed at once.