#21 feat: started redesigning the app
Added profile header to the Home page. Redesigned the add button, the searchbar, the list filters and listOrderModal. For some reason i had to add the Statusbar in the _layout.tsx because it was hidden.

note: removed the vertical dots from the lists to chevrons for further updates.

chore: added async-storage and updated expo dependencies and expo SDK 52